Handle NICK/USER messages in separate step
previously we would add "partial" users to the list of users when receiving
NICK/USER messages, and use hostname as the key to find the users. Now
we have a separate while loop that will read only NICK/USER messages
until a valid user struct (ie all fields are complete and no conflicting
nickname) is constructed. Only at that point will it register the user and
from then on the nick should be used as the key for each user.