Rewrite reply writing
Have each msg handler send its own replies instead of accepting and writing into
a buffer and having the caller send the reply, which simplifies the code a little.

Also reimplement the logic for writing replies. Having a function which takes in
a list of arguments and then writes them into a string is unecessary - it's a lot
easier to just have the caller sprintf its own args into a string. Then just have
a function to add the reply prefix and send the message.